refactor: #123 - resend-verification-email mutation → mutations 디렉터리명 통일 및 import 경로 정리
